finish

+++ TasksManagement/Controllers/TasksController.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -9,23 +12,31 @@ using TasksManagement.Models;
 
 namespace TasksManagement.Controllers
 {
-    //[Authorize]
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class TasksController : ControllerBase
     {
         private readonly IUnitOfWork _tasksUnitOfWork;
+        readonly UserManager<Data.Entities.User> _userManager;
 
-        public TasksController(IUnitOfWork tasksUnitOfWork)
+        public TasksController(IUnitOfWork tasksUnitOfWork, UserManager<Data.Entities.User> userManager)
         {
             _tasksUnitOfWork = tasksUnitOfWork;
+            _userManager = userManager;
         }
 
         // GET: api/Tasks
         [HttpGet]
-        public IEnumerable<TaskModel> GetTasks()
+        public async Task<IEnumerable<TaskModel>> GetTasks()
         {
-            return from task in _tasksUnitOfWork.GetTaksRepository().GetAll().IncludeMultiple(t => t.OwnerUser, t => t.AssignedToUser)
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+            var roles = await _userManager.GetRolesAsync(user);
+
+            string role = roles.FirstOrDefault();
+
+            return from task in _tasksUnitOfWork.GetTaksRepository().GetAll().Where(p => p.OwnerUserID == user.Id || p.AssignedToUserID == user.Id || role.ToLower() == "admin")
+                   .IncludeMultiple(t => t.OwnerUser, t => t.AssignedToUser)
                    select new TaskModel
                    {
                        ID = task.ID,
@@ -34,16 +45,17 @@ namespace TasksManagement.Controllers
                        Created = task.Created,
                        OwnerUser = new UserModel()
                        {
-                           UserID = task.OwnerUser.Id,
+                           ID = task.OwnerUserID,
                            Username = task.OwnerUser.UserName,
                            Email = task.OwnerUser.Email
                        },
                        AssignedToUser = task.AssignedToUser != null ? new UserModel()
                        {
-                           UserID = task.AssignedToUser.Id,
+                           ID = task.AssignedToUserID,
                            Username = task.AssignedToUser.UserName,
                            Email = task.AssignedToUser.Email
                        } : null,
+                       Completed = task.Completed
                    };
         }
 
@@ -68,16 +80,47 @@ namespace TasksManagement.Controllers
 
         // PUT: api/Tasks/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutTask([FromRoute] int id, [FromBody] Data.Entities.Task task)
+        public async Task<IActionResult> PutTask([FromRoute] int id, [FromBody] TaskModel model)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            if (id != task.ID)
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+
+            var tasks = await _tasksUnitOfWork.GetTaksRepository().GetAsync(p => p.ID == id);
+            if (tasks.Count == 0)
+                return NotFound();
+
+            var task = tasks.FirstOrDefault();
+            if (task == null)
+                return NotFound();
+
+            if (model.AssignedToUser == null)
             {
-                return BadRequest();
+                if (model.Completed)
+                {
+                    if (task.AssignedToUserID == user.Id)
+                    {
+                        task.Completed = true;
+                        task.AssignedToUserID = null;
+                    }
+                    else
+                        return BadRequest(new { message = "This task is not assigned to you" });
+                }
+            }
+            else
+            {
+                if (task.OwnerUserID == user.Id)
+                {
+                    if (model.AssignedToUser.ID == user.Id)
+                        return BadRequest(new { message = "You can't assign the task to yourself" });
+                    else
+                        task.AssignedToUserID = model.AssignedToUser.ID;
+                }
+                else
+                    return BadRequest(new { message = "You are not owner for this task" });
             }
 
             _tasksUnitOfWork.GetTaksRepository().Edit(task);
@@ -98,45 +141,84 @@ namespace TasksManagement.Controllers
                 }
             }
 
-            return NoContent();
+            return Ok(new TaskModel
+            {
+                ID = task.ID,
+                Title = task.Title,
+                Description = task.Description,
+                Created = task.Created,
+                OwnerUser = new UserModel()
+                {
+                    ID = task.OwnerUserID,
+                    Username = task.OwnerUser.UserName,
+                    Email = task.OwnerUser.Email
+                },
+                AssignedToUser = task.AssignedToUser != null ? new UserModel()
+                {
+                    ID = task.AssignedToUserID,
+                    Username = task.AssignedToUser.UserName,
+                    Email = task.AssignedToUser.Email
+                } : null,
+                Completed = task.Completed
+            });
         }
 
         // POST: api/Tasks
         [HttpPost]
-        public async Task<IActionResult> PostTask([FromBody] Data.Entities.Task task)
+        public async Task<IActionResult> PostTask([FromBody] TaskModel task)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            _tasksUnitOfWork.GetTaksRepository().Add(task);
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+
+            _tasksUnitOfWork.GetTaksRepository().Add(new Data.Entities.Task()
+            {
+                Title = task.Title,
+                Description = task.Description,
+                OwnerUserID = user.Id,
+                Created = DateTime.Now
+            });
 
             await _tasksUnitOfWork.SaveChangesAsync();
 
-            return CreatedAtAction("GetTask", new { id = task.ID }, task);
+            return RedirectToAction("GetTasks");
         }
 
         // DELETE: api/Tasks/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteTask([FromRoute] int id)
         {
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+            var roles = await _userManager.GetRolesAsync(user);
+
+            string role = roles.FirstOrDefault();
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            var task = await _tasksUnitOfWork.GetTaksRepository().GetAsync(p => p.ID == id);
-            if (task == null)
-            {
+            var tasks = await _tasksUnitOfWork.GetTaksRepository().GetAsync(p => p.ID == id);
+            if (tasks.Count == 0)
                 return NotFound();
+
+            var task = tasks.FirstOrDefault();
+            if (task == null)
+                return NotFound();
+
+            if ((role.ToLower() == "admin") || (task.OwnerUserID == user.Id))
+            {
+                _tasksUnitOfWork.GetTaksRepository().Delete(p => p.ID == id);
+
+                await _tasksUnitOfWork.SaveChangesAsync();
+
+                return Ok(task);
             }
 
-            _tasksUnitOfWork.GetTaksRepository().Delete(p => p.ID == id);
-
-            await _tasksUnitOfWork.SaveChangesAsync();
-
-            return Ok(task);
+            return BadRequest(new { message = "You are not owner or admin to delete this task" });
         }
 
         private bool TaskExists(int id)
